+++ common/src/main/java/net/frozenblock/lib/block/api/dripstone/DripstoneDripApi.java
@@ -22,7 +22,8 @@ import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import net.frozenblock.lib.block.api.tick.BlockScheduledTicks;
+import lombok.experimental.UtilityClass;
+import net.frozenblock.lib.block.api.tick.BlockTickEvents;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.block.Block;
@@ -34,9 +35,10 @@ import org.jetbrains.annotations.ApiStatus;
 /**
  * An API related to Pointed Dripstone dripping.
  * <p>
- * If you are looking to add a custom funciton when a block is dripped ON, it is recommended to use {@link BlockScheduledTicks} and {@link #getDripstoneFluid(ServerLevel, BlockPos)}.
+ * If you are looking to add a custom funciton when a block is dripped ON, it is recommended to use {@link BlockTickEvents} and {@link #getDripstoneFluid(ServerLevel, BlockPos)}.
  */
-public class DripstoneDripApi {
+@UtilityClass
+public final class DripstoneDripApi {
 	private static final Map<Block, List<InjectedDrip>> WATER_DRIP_METHODS = new Object2ObjectOpenHashMap<>();
 	private static final Map<Block, List<InjectedDrip>> LAVA_DRIP_METHODS = new Object2ObjectOpenHashMap<>();