+++ common/src/main/java/net/frozenblock/lib/block/mixin/tick/BlockStateBaseMixin.java
@@ -19,8 +19,7 @@ package net.frozenblock.lib.block.mixin.tick;
 
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
-import net.frozenblock.lib.block.api.tick.BlockRandomTicks;
-import net.frozenblock.lib.block.api.tick.BlockScheduledTicks;
+import net.frozenblock.lib.block.api.tick.BlockTickEvents;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.util.RandomSource;
@@ -40,8 +39,8 @@ public class BlockStateBaseMixin {
 			target = "Lnet/minecraft/world/level/block/Block;tick(Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/server/level/ServerLevel;Lnet/minecraft/core/BlockPos;Lnet/minecraft/util/RandomSource;)V"
 		)
 	)
-	public void frozenLib$runCustomTick(Block instance, BlockState state, ServerLevel level, BlockPos pos, RandomSource random, Operation<Void> original) {
-		BlockScheduledTicks.runTickIfPresent(state, level, pos, random);
+	public void frozenLib$invokeTick(Block instance, BlockState state, ServerLevel level, BlockPos pos, RandomSource random, Operation<Void> original) {
+		BlockTickEvents.TICK.invoker().onTick(state, level, pos, random);
 		original.call(instance, state, level, pos, random);
 	}
 
@@ -52,8 +51,8 @@ public class BlockStateBaseMixin {
 			target = "Lnet/minecraft/world/level/block/Block;randomTick(Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/server/level/ServerLevel;Lnet/minecraft/core/BlockPos;Lnet/minecraft/util/RandomSource;)V"
 		)
 	)
-	public void frozenLib$runCustomRandomTick(Block instance, BlockState state, ServerLevel level, BlockPos pos, RandomSource random, Operation<Void> original) {
-		BlockRandomTicks.runRandomTickIfPresent(state, level, pos, random);
+	public void frozenLib$invokeRandomTick(Block instance, BlockState state, ServerLevel level, BlockPos pos, RandomSource random, Operation<Void> original) {
+		BlockTickEvents.RANDOM_TICK.invoker().onRandomTick(state, level, pos, random);
 		original.call(instance, state, level, pos, random);
 	}
 }